DevSecOpsKB-router-query-engine-document-management/requirements.txt
python-dotenv
openai
langchain
llama-index==0.6.38
ndg-httpsclient
pypdf
PyCryptodome
gradio